migrate the method that get user from authorization header to S3RestUtils.
