+++ core/server/proxy/src/test/java/alluxio/proxy/s3/S3RestServiceHandlerTest.java
@@ -28,18 +28,18 @@ public class S3RestServiceHandlerTest {
   @Test
   public void userFromAuthorization() throws Exception {
     try {
-      S3AuthenticationFilter.getUserFromAuthorization("AWS-SHA256-HMAC Credential=/asd", mConf);
+      S3RestUtils.getUserFromAuthorization("AWS-SHA256-HMAC Credential=/asd", mConf);
       Assert.fail();
     } catch (Exception e) {
       Assert.assertTrue(e instanceof S3Exception);
     }
     try {
-      assertNull(S3AuthenticationFilter.getUserFromAuthorization("", mConf));
+      assertNull(S3RestUtils.getUserFromAuthorization("", mConf));
       Assert.fail();
     } catch (Exception e) {
       Assert.assertTrue(e instanceof S3Exception);
     }
-    assertEquals("test", S3AuthenticationFilter.getUserFromAuthorization(
+    assertEquals("test", S3RestUtils.getUserFromAuthorization(
         "AWS-SHA256-HMAC Credential=test/asd", mConf));
   }
